requirements.txt
bs4
yaspin
pysocks
requests
streamlit
python-dotenv
langchain-openai
langchain-ollama
langchain-anthropic
langchain_community
langchain_google_genai
urllib3>=1.26.0
stem
telethon
duckduckgo-search
reportlab
fastapi
uvicorn[standard]
slowapi